fix: phase2.3 hardening
- Price/Cost: float64 → int64 (thousandths of a cent, stored as BIGINT in DB)
- Removed PreviousStatus from Message struct (events + audit logs cover tracking)
- Migration: NUMERIC(12,4) → BIGINT for price/cost, removed previous_status column
- Composite indexes: (tenant_id, status, created_at DESC), (tenant_id, connector_id), (tenant_id, direction)
- Documented outbox pattern as 'publish-after-commit' (not true outbox yet)
- Added note in service to clarify this is publish-after-commit pattern

+++ internal/service/message_service.go
@@ -149,7 +149,7 @@ func (s *MessageService) SendMessage(ctx context.Context, id string, connectorID
 }
 
 // MarkSent transitions a message from sending → sent.
-func (s *MessageService) MarkSent(ctx context.Context, id, externalID string, parts int, price, cost float64) (*domain.Message, error) {
+func (s *MessageService) MarkSent(ctx context.Context, id, externalID string, parts int, price, cost int64) (*domain.Message, error) {
 	status := domain.MessageStatusSent
 	now := s.clock.Now()
 	return s.UpdateStatus(ctx, id, domain.UpdateMessageInput{
